// src/app/api/generated/model/update-arancel-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Cambios a aplicar sobre un arancel. Los campos nulos no se tocan
 */
export interface UpdateArancelRequest { 
    /**
     * La parte que paga el paciente
     */
    coseguro?: number;
    /**
     * Version leida del arancel
     */
    expectedVersion: number;
    /**
     * La parte que paga el financiador
     */
    importeFinanciador?: number;
    /**
     * Lo que vale la practica bajo este convenio
     */
    importeTotal?: number;
    /**
     * Primer dia en que este importe se aplica
     */
    vigenciaDesde?: string;
    /**
     * Ultimo dia en que se aplica, INCLUSIVE
     */
    vigenciaHasta?: string;
}